Check for duplicate properties in ATL
This adds two concurrent checks, which is pointless, but : it only checks for duplicate properties in the same warper (so on a single line by themselves, or in a `linear 5:` block). It needs to also trigger for two `xpos 5` alone in successive lines with no warper, and I'm not sure which check will do it best. So both are here.