Merge consecutive properties statements
Caveat : the error doesn't show the proper location in the code. That check should be done in the while loop earlier.
Also, the is_properties is too messy, there should be a better, cleaner way to do this. Maybe by making a specialized class separate from MultiPurpose for simple properties ?